pyproject.toml
[project]
name = "discord-python"
version = "0.2.0"
description = "A simple Discord bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
package = false